Ajuste Final

+++ src/app/api/clientes/[id]/contactos/[id_contacto]/route.tsx
@@ -1,17 +1,27 @@
-import { PrismaClient } from '@prisma/client'
-import { NextResponse } from 'next/server'
+import { PrismaClient } from '@prisma/client';
+import { NextRequest, NextResponse } from 'next/server';
 
-const prisma = new PrismaClient()
+const prisma = new PrismaClient();
 
-export async function DELETE(_: Request, { params }: { params: { id: string, id_contacto: string } }) {
+export async function DELETE(req: NextRequest) {
+  const url = req.nextUrl;
+  const pathSegments = url.pathname.split('/');
+
+  const id = pathSegments[pathSegments.indexOf('clientes') + 1];
+  const id_contacto = pathSegments[pathSegments.indexOf('contactos') + 1];
+
+  if (!id || !id_contacto) {
+    return NextResponse.json({ error: 'Parámetros inválidos' }, { status: 400 });
+  }
+
   await prisma.listas_contactos_contactos.delete({
     where: {
       id_lista_id_contacto: {
-        id_lista: parseInt(params.id),
-        id_contacto: parseInt(params.id_contacto),
+        id_lista: parseInt(id),
+        id_contacto: parseInt(id_contacto),
       }
     }
-  })
+  });
 
-  return NextResponse.json({ ok: true })
+  return NextResponse.json({ ok: true });
 }
